wdk_mutex::kmutex
Struct KMutex

pub struct KMutex<T> { /* private fields */ }

A thread safe mutex implemented through acquiring a KMUTEX in the Windows kernel.

The type Kmutex<T> provides mutually exclusive access to the inner type T allocated through this crate in the non-paged pool. All data required to initialise the KMutex is allocated in the non-paged pool and as such is safe to pass stack data into the type as it will not go out of scope.

KMutex holds an inner value which is a pointer to a KMutexInner type which is the actual type allocated in the non-paged pool, and this holds information relating to the mutex.

Access to the T within the KMutex can be done through calling Self::lock.

To receive debug messages when the IRQL is too high for an operation, enable the feature flag debug.

Lifetimes

As the KMutex is designed to be used in the Windows Kernel, with the Windows wdk crate, the lifetimes of the KMutex must be considered by the caller. See examples below for usage.

The KMutex can exist in a locally scoped function with little additional configuration. To use the mutex across thread boundaries, or to use it in callback functions, the recommended course of action is to utilise either a globally accessible static AtomicPtr<KMutex<T>>; or to utilise a Device Extension provided in the wdk.

If you use a `static AtomicPtr>` you MUST ensure that the memory is cleaned up when you exit the driver otherwise you cause a memory leak.
A future addition is planned which will make the API more flexible for dynamically managing globally available mutexes to somewhat reduce the overhead required to use this crate.

Deallocation

KMutex handles the deallocation of resources at the point the KMutex is dropped.

Examples

Locally scoped mutex:

{
    let mtx = KMutex::new(0u32).unwrap();
    let lock = mtx.lock().unwrap();
 
    // If T implements display, you do not need to dereference the lock to print.
    println!("The value is: {}", lock);
} // Mutex will become unlocked as it is managed via RAII 

Global scope via static pointer:

A future release is planned to make this process more ergonomic.

pub static HEAP_MTX_PTR: AtomicPtr<KMutex<u32>> = AtomicPtr::new(null_mut());
 
fn my_fn() {
    let heap_mtx = Box::new(KMutex::new(0u32).unwrap());
    let heap_mtx_ptr = Box::into_raw(heap_mtx);
    HEAP_MTX_PTR.store(heap_mtx_ptr, Ordering::SeqCst);
 
    // spawn some system threads
    r _ in 0..3 {
        let mut thread_handle: HANDLE = null_mut();
        let status = unsafe {
            PsCreateSystemThread(
                &mut thread_handle, 
                0, 
                null_mut::<OBJECT_ATTRIBUTES>(), 
                null_mut(),
                null_mut::<CLIENT_ID>(), 
                Some(callback_fn), 
                null_mut(),
            )
        };
        println!("[i] Thread status: {status}");
    }
}
 
unsafe extern "C" fn callback_fn(_: *mut c_void) {
    for _ in 0..50 {
        let p = HEAP_MTX_PTR.load(Ordering::SeqCst);
        if !p.is_null() {
            let p = unsafe { &*p };
            let mut lock = p.lock().unwrap();
            println!("Got the lock before change! {}", *lock);
            *lock += 1;
            println!("After the change: {}", *lock);
        }
    }
}
 
// IMPORTANT ensure the KMutex in the static is properly dropped to clean memory
extern "C" fn driver_exit(driver: *mut DRIVER_OBJECT) {
    let ptr: *mut KMutex<u32> = HEAP_MTX_PTR.load(Ordering::SeqCst);
    if !ptr.is_null() {
        unsafe {
            // RAII will kick in here to deallocate our memory
            let _ = Box::from_raw(ptr);
        }
    }
}

Implementations

impl<T> KMutex<T>

pub fn new(data: T) -> Result<Self, DriverMutexError>

Creates a new KMUTEX Windows Kernel Driver Mutex in a signaled (free) state.

IRQL

This can be called at any IRQL.

Examples

use wdk_mutex::Mutex;
 
let my_mutex = wdk_mutex::KMutex::new(0u32);

pub fn lock(&self) -> Result<KMutexGuard<'_, T>, DriverMutexError>

Acquires a mutex in a non-alertable manner.

A future release is planned to include an alternate implementation which will lock the mutex and become alertable if it has to wait for the mutex to become free. This function will block the local thread until it is available to acquire the mutex.

Once the thread has acquired the mutex, it will return a KMutexGuard which is a RAII scoped guard allowing exclusive access to the inner T.

Errors

If the IRQL is too high, this function will return an error and will not acquire a lock. To prevent a kernel panic, the caller should match the return value rather than just unwrapping the value.

IRQL

This function must be called at IRQL <= APC_LEVEL, if the IRQL is higher than this, the function will return an error.

It is the callers responsibility to ensure the IRQL is sufficient to call this function and it will not alter the IRQL for the caller, as this may introduce undefined behaviour elsewhere in the driver / kernel.

Examples

let mtx = KMutex::new(0u32).unwrap();
let lock = mtx.lock().unwrap();